refactor: apis, hooks 함수별로 분리
