+++ src/services/lunch/apis/getLunchMenusWithPollStatus.ts
@@ -2,9 +2,9 @@ import type { LunchMenusWithPollStatus } from '~/services/lunch/utils';
 import type { ApiSuccessResponse } from '~/types';
 
 import { endpoints } from '~/react-query/common';
-import { privateAxios } from '~/utils';
+import { privateAxios } from '~/utils/axios';
 
-export interface GetLunchMenusWithPollStatus {
+export interface GetLunchMenusWithPollStatusParams {
   campus: string;
   date: string;
 }
@@ -13,7 +13,7 @@ export type GetLunchMenusWithPollStatusApiData =
   ApiSuccessResponse<LunchMenusWithPollStatus>;
 
 export const getLunchMenusWithPollStatus = (
-  params: GetLunchMenusWithPollStatus
+  params: GetLunchMenusWithPollStatusParams
 ) => {
   const { campus, date } = params;
 
@@ -22,16 +22,4 @@ export const getLunchMenusWithPollStatus = (
   return privateAxios
     .get<GetLunchMenusWithPollStatusApiData>(endpoint)
     .then((res) => res.data.data);
-};
-
-export const pollLunchMenu = (lunchId: number) => {
-  const endpoint = endpoints.lunch.vote(lunchId);
-
-  return privateAxios.post(endpoint).then((res) => res.data);
-};
-
-export const revertPolledLunchMenu = (lunchId: number) => {
-  const endpoint = endpoints.lunch.revertVote(lunchId);
-
-  return privateAxios.post(endpoint).then((res) => res.data);
 };
