+++ src/services/lunch/apis/pollLunchMenu.ts
@@ -1,0 +1,8 @@
+import { endpoints } from '~/react-query/common';
+import { privateAxios } from '~/utils/axios';
+
+export const pollLunchMenu = (lunchId: number) => {
+  const endpoint = endpoints.lunch.vote(lunchId);
+
+  return privateAxios.post(endpoint).then((res) => res.data);
+};
